+++ scripts/background.js
@@ -138,8 +138,6 @@ function sendTab(element) {
 }
 
 function logit(message) {
-	//var options = { day: '2-digit', year: 'numeric', month: '2-digit', hour: '2-digit', minute: '2-digit', second: '2-digit' };
-	//var mDate = new Date().toLocaleString(undefined, options);
 	var ndate = new Date();
 	logline = loglines + ndate.toLocaleString() + " - " + message + "\n";
 	if(message.toString().toLowerCase().indexOf('error') >= 0 && message.toString().toLowerCase().indexOf('TypeError') <= 0) 
@@ -148,6 +146,11 @@ function logit(message) {
 }
 
 function init() {
+	loglines = logit("Info: AddOn version: "+chrome.runtime.getManifest().version);
+	chrome.runtime.getPlatformInfo(function(info){
+		loglines = logit("Info: Current architecture: "+info.arch+" | Current OS: "+info.os);
+	});
+	loglines = logit("Info: "+navigator.userAgent);
 	chrome.storage.local.set({last_message: ""});
 	chrome.storage.local.get(null, function(options) {
 		var s_startup = options['s_startup'] || false;
@@ -179,7 +182,7 @@ function getNotifications() {
 			if( xhr.status < 200 || xhr.status > 226) {
 				message = "Get list of notifications failed.";
 				notify('error',message);
-				loglines = logit('Error: ' + message);
+				loglines = logit('Error: '+message);
 			}
 			else {
 				if(xhr.responseText.length > 0) {
@@ -433,7 +436,7 @@ function exportPHPMarks() {
 					}
 				}
 			}
-			loglines = logit("Info: Sending export of local bookmarks for client "+s_uuid);
+			loglines = logit("Info: Sending export of local bookmarks to server");
 			xhr.send(cdata);
 		})
 	});
@@ -511,7 +514,7 @@ function delMark(id, bookmark) {
 					}
 			}
 		}
-		loglines = logit("Info: Sending remove request to server. URL: "+bookmark.node.url+", Client: "+s_uuid);
+		loglines = logit("Info: Sending remove request to server: <a href='"+bookmark.node.url+"'>"+bookmark.node.url+"</a>");
 		xhr.send(cdata);
 	})
 	
@@ -551,10 +554,10 @@ function moveMark(id, bookmark) {
 						if(response == 1)
 								loglines = logit("Info: Bookmark moved successfully at the server");
 							else
-								loglines = logit("Error: Bookmark not moved at the server, response from server is: " + response);
+								loglines = logit("Error: Bookmark not moved at the server, response from server is: "+response);
 					}
 				}
-				loglines = logit("Info: Sending move request to server. Bookmark ID: "+id+", Client: "+s_uuid);
+				loglines = logit("Info: Sending move request to server. Bookmark ID: "+id);
 				xhr.send(cdata);
 			});
 		});
@@ -613,10 +616,10 @@ function sendMark(bookmark) {
 					if(response == 1)
 							loglines = logit("Info: Bookmark added successfully at the server");
 						else
-							loglines = logit("Error: Bookmark not added at the server, please check the server logfile");
+							loglines = logit("Error: Bookmark not added at the server. Response: "+response);
 				}
 			}
-			loglines = logit("Info: Sending add request to server. URL: "+bookmark.url+", Client: "+s_uuid);
+			loglines = logit("Info: Sending add request to server. <a href='"+bookmark.url+"'>"+bookmark.url+"</a>");
 			xhr.send(cdata);
 		});
 	});
@@ -739,7 +742,7 @@ function getAllPHPMarks() {
 			}
 			chrome.browserAction.setTitle({title: chrome.i18n.getMessage("extensionName") + ": " + date.toLocaleDateString(undefined,doptions)});
 			}
-		loglines = logit('Info: Sending import request to server. Client: '+options['s_uuid']);
+		loglines = logit('Info: Sending import request to server');
 		xhr.send(params);
 	});
 }
@@ -755,14 +758,13 @@ function c2cm(bookmarks) {
 
 function addPHPcMarks(bArray) {
 	bArray.forEach(function(bookmark) {
-		if(bookmark.bmAction == 1 && bookmark.bmURL != '') {
-			loglines = logit('Info: Try to remove bookmark '+bookmark.bmURL);
+		if(bookmark.bmAction == 1 && bookmark.bmURL.length > 0) {
+			loglines = logit("Info: Try to remove bookmark <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a>");
 			chrome.bookmarks.search({url: bookmark.bmURL}, function(removeItems) {
 				removeItems.forEach(function(removeBookmark) {
 					if(removeBookmark.dateAdded == bookmark.bmAdded) {
 						chrome.bookmarks.onRemoved.removeListener(onRemovedCheck);
 						chrome.bookmarks.remove(removeBookmark.id, function(remove) {
-							loglines = logit('Info: Bookmark '+removeBookmark.url+' removed');
 							chrome.bookmarks.onRemoved.addListener(onRemovedCheck);
 						});
 					}
@@ -771,7 +773,7 @@ function addPHPcMarks(bArray) {
 		}
 		else {
 			if(bookmark.fdID.length > 1) {
-				loglines = logit('Info: Changed bookmark '+bookmark.bmURL+' in userfolder');
+				loglines = logit("Info: Changed bookmark <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a> in userfolder");
 				chrome.bookmarks.search({title: bookmark.fdName},function(folderItems) {
 					folderItems.forEach(function(folder) {
 						if(folder.index == bookmark.fdIndex) {
@@ -780,14 +782,14 @@ function addPHPcMarks(bArray) {
 									if (bookmark.fdName != bookmarkItems[0].parentId) {
 										chrome.bookmarks.onMoved.removeListener(onMovedCheck);
 										chrome.bookmarks.move(bookmarkItems[0].id, {parentId: folder.id}, function(move) {
-											loglines = logit('Info: '+move.url + " moved to folder " + folder.title);
+											loglines = logit("Info: <a href='"+move.url+"'>"+move.url+"</a> moved to folder " + folder.title);
 											chrome.bookmarks.onMoved.addListener(onMovedCheck);
 										});
 									}
 								} else {
 									chrome.bookmarks.onCreated.removeListener(onCreatedCheck);
 									chrome.bookmarks.create({parentId: folder.id, title: bookmark.bmTitle, url: bookmark.bmURL }, function() {
-										loglines = logit('Info: '+bookmark.bmURL + " added as new bookmark");
+										loglines = logit("Info: <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a> added as new bookmark");
 										chrome.bookmarks.onCreated.addListener(onCreatedCheck);
 									});
 								}
@@ -796,7 +798,7 @@ function addPHPcMarks(bArray) {
 						} else {
 							chrome.bookmarks.onCreated.removeListener(onCreatedCheck);
 							chrome.bookmarks.create({parentId: folder.id, title: bookmark.bmTitle, url: bookmark.bmURL }, function() {
-								loglines = logit('Info: '+bookmark.bmURL+" added as new bookmark");
+								loglines = logit("Info: <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a> added as new bookmark");
 								chrome.bookmarks.onCreated.addListener(onCreatedCheck);
 							});
 						}
@@ -804,22 +806,21 @@ function addPHPcMarks(bArray) {
 				});
 			}
 			else {
-				loglines = logit('Info: Changed bookmark '+bookmark.bmURL+' in systemfolder');
-				if(bookmark.bmURL != '') {
-					loglines = logit('Info: Try to add bookmark '+bookmark.bmURL);
+				if(bookmark.bmURL != null) {
+					loglines = logit("Info: Try to add bookmark <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a>");
 					chrome.bookmarks.search({url: bookmark.bmURL}, function(bookmarkItems) {
 						if (bookmarkItems.length) {
 							if(bookmarkItems[0].parentId != bookmark.fdID) {
 								chrome.bookmarks.onMoved.removeListener(onMovedCheck);
 								chrome.bookmarks.move(bookmarkItems[0].id, {parentId: bookmark.fdID}, function(move) {
-									loglines = logit('Info: '+move.url + " moved to folder " + bookmark.fdName);
+									loglines = logit("Info: <a href='"+move.url+"'>"+move.url+"</a> moved to folder " + bookmark.fdName);
 									chrome.bookmarks.onMoved.addListener(onMovedCheck);
 								});
 							}
 						} else {
 							chrome.bookmarks.onCreated.removeListener(onCreatedCheck);
 							chrome.bookmarks.create({parentId: bookmark.fdID, title: bookmark.bmTitle, url: bookmark.bmURL}, function() {
-								loglines = logit('Info: '+bookmark.bmURL + " added as new bookmark.");
+								loglines = logit("Info: <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a> added as new bookmark.");
 								chrome.bookmarks.onCreated.addListener(onCreatedCheck);
 							});
 						}
@@ -832,14 +833,14 @@ function addPHPcMarks(bArray) {
 
 function addPHPMarks(bArray) {
 	bArray.forEach(function(bookmark) {
-		if(bookmark.bmAction == 1 && bookmark.bmURL != '') {
-			loglines = logit('Info: Try to remove bookmark '+bookmark.bmURL);
+		if(bookmark.bmAction == 1 && bookmark.bmURL != null) {
+			loglines = logit("Info: Try to remove bookmark <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a>");
 			chrome.bookmarks.search({url: bookmark.bmURL}, function(removeItems) {
 				removeItems.forEach(function(removeBookmark) {
 					if(removeBookmark.dateAdded == bookmark.bmAdded) {
 						chrome.bookmarks.onRemoved.removeListener(onRemovedCheck);
 						chrome.bookmarks.remove(removeBookmark.id, function(remove) {
-							loglines = logit('Info: Bookmark '+removeBookmark.url+' removed');
+							loglines = logit("Info: Bookmark <a href='"+removeBookmark.url+"'>"+removeBookmark.url+"</a> removed");
 							chrome.bookmarks.onRemoved.addListener(onRemovedCheck);
 						});
 					}
@@ -848,7 +849,7 @@ function addPHPMarks(bArray) {
 		}
 		else {
 			if(!bookmark.fdID.endsWith('___')) {
-				loglines = logit('Info: Changed bookmark '+bookmark.bmURL+' in userfolder');
+				loglines = logit("Info: Changed bookmark <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a> in userfolder");
 				chrome.bookmarks.search({title: bookmark.fdName},function(folderItems) {
 					folderItems.forEach(function(folder) {
 						if(folder.index == bookmark.fdIndex) {
@@ -857,14 +858,14 @@ function addPHPMarks(bArray) {
 									if (bookmark.fdName != bookmarkItems[0].parentId) {
 										chrome.bookmarks.onMoved.removeListener(onMovedCheck);
 										chrome.bookmarks.move(bookmarkItems[0].id, {parentId: folder.id}, function(move) {
-											loglines = logit('Info: '+move.url + " moved to folder " + folder.title);
+											loglines = logit("Info: <a href='"+move.url+"'>"+move.url+"</a> moved to folder " + folder.title);
 											chrome.bookmarks.onMoved.addListener(onMovedCheck);
 										});
 									}
 								} else {
 									chrome.bookmarks.onCreated.removeListener(onCreatedCheck);
 									chrome.bookmarks.create({type: bookmark.bmType, parentId: folder.id, title: bookmark.bmTitle, url: bookmark.bmURL }, function() {
-										loglines = logit('Info: '+bookmark.bmURL + " added as new bookmark");
+										loglines = logit("Info: <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a> added as new bookmark");
 										chrome.bookmarks.onCreated.addListener(onCreatedCheck);
 									});
 								}
@@ -873,7 +874,7 @@ function addPHPMarks(bArray) {
 						} else {
 							chrome.bookmarks.onCreated.removeListener(onCreatedCheck);
 							chrome.bookmarks.create({type: bookmark.bmType, parentId: folder.id, title: bookmark.bmTitle, url: bookmark.bmURL }, function() {
-								loglines = logit('Info: '+bookmark.bmURL+" added as new bookmark");
+								loglines = logit("Info: <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a> added as new bookmark");
 								chrome.bookmarks.onCreated.addListener(onCreatedCheck);
 							});
 						}
@@ -881,22 +882,21 @@ function addPHPMarks(bArray) {
 				});
 			}
 			else {
-				loglines = logit('Info: Changed bookmark '+bookmark.bmURL+' is in systemfolder');
-				if(bookmark.bmURL != '') {
+				if(bookmark.bmURL != null) {
 					loglines = logit('Info: Try to add bookmark '+bookmark.bmURL);
 					chrome.bookmarks.search({url: bookmark.bmURL}, function(bookmarkItems) {
 						if (bookmarkItems.length) {
 							if(bookmarkItems[0].parentId != bookmark.fdID) {
 								chrome.bookmarks.onMoved.removeListener(onMovedCheck);
 								chrome.bookmarks.move(bookmarkItems[0].id, {parentId: bookmark.fdID}, function(move) {
-									loglines = logit('Info: '+move.url + " moved to folder " + bookmark.fdName);
+									loglines = logit("Info: <a href='"+move.url+"'>"+move.url+"</a> moved to folder " + bookmark.fdName);
 									chrome.bookmarks.onMoved.addListener(onMovedCheck);
 								});
 							}
 						} else {
 							chrome.bookmarks.onCreated.removeListener(onCreatedCheck);
 							chrome.bookmarks.create({type: bookmark.bmType, parentId: bookmark.fdID, title: bookmark.bmTitle, url: bookmark.bmURL}, function() {
-								loglines = logit('Info: '+bookmark.bmURL + " added as new bookmark.");
+								loglines = logit("Info: <a href='"+bookmark.bmURL+"'>"+bookmark.bmURL+"</a> added as new bookmark.");
 								chrome.bookmarks.onCreated.addListener(onCreatedCheck);
 							});
 						}